requirements.txt
streamlit
ultralytics
opencv-python
torch
torchvision
pillow
numpy

